==> 1_setup/species/iucn/2_generate_spp_map_list.Rmd <==
---
title: 'Generate IUCN spp map list'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '~/github/src/templates/ohara_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'figs/',
                      echo = TRUE, message = FALSE, warning = FALSE)

library(raster)
source('https://raw.githubusercontent.com/oharac/src/master/R/common.R')

source(here::here('common_fxns.R'))
source('iucn_api_fxns.R') ### in same folder as this script

dir_bli <- '/home/shares/ohi/git-annex/globalprep/_raw_data/birdlife_intl/d2021'
dir_shp <- '/home/shares/ohi/git-annex/globalprep/_raw_data/iucn_spp/d2021-3'
  ### in this folder are shapefiles at a taxonomic level, from
  ### IUCN spatial data download (version 2021-3).

```

# Summary

* Subpopulation polygons must be identified and rasterized separately from the parent polygon; this must be done by sciname and subpop fields since the polygon IDs are based upon the parent ID.
* Regional assessments need not be determined at this stage - the ID numbers match the global ID numbers (including subpops).

# Data sources

* IUCN species shapefiles:  IUCN. (2019). The IUCN Red List of Threatened Species. Version 2019-2. Retrieved August 2019, from http://www.iucnredlist.org
* BirdLife International shapefiles: BirdLife International and Handbook of the Birds of the World. (2018). Bird species distribution maps of the world. Version 7.0.  Available at http://datazone.birdlife.org/species/requestdis


# Methods

## Generate species map list with subpops

### Create subpopulation lookup to match maps with IUCN SIDs for subpops

From available shapefiles, reclassify subpopulation polygons into subpop ID numbers for matching to risk and trend. 

``` {r create_subpop_lookup}

iucn_dbfs <- list.files(dir_shp, pattern = '\\.dbf$', full.names = TRUE)
bli_dbfs  <- list.files(dir_bli, pattern = '\\.dbf$', full.names = TRUE)
shps_dbfs <- c(iucn_dbfs, bli_dbfs)

### This is all "marine" species by habitat and system:
marine_spp <- read_csv(here('_data/iucn_spp', 
                            sprintf('spp_marine_from_api_%s.csv', api_version)))

### This is *all* spp from the API, filtered down to those in marine_spp: 
api_spp <- read_csv(here_anx('iucn_spp', sprintf('spp_info_from_api_%s.csv', api_version))) %>%
  filter(iucn_sid %in% marine_spp$iucn_sid)

### Define helper function to cleanly get the info from DBF files,
### including colname checks
get_dbf <- function(x) { ### x <- shps_dbfs[1]
  spp_dbf_info <- foreign::read.dbf(x, as.is = TRUE) %>%
    clean_df_names()
  
  ### Downloaded files have id_no; BirdLife uses sisid
  names(spp_dbf_info)[names(spp_dbf_info) %in% c('id_no', 'sisid')] <- 'iucn_sid'

  ### if 'binomial' ('binomil' in BLI), set to 'sciname'
  names(spp_dbf_info)[str_detect(names(spp_dbf_info), '^binom')] <- 'sciname'

  ### fix BLI 'presenc' 
  names(spp_dbf_info)[str_detect(names(spp_dbf_info), '^presenc')] <- 'presence'

  ### Some files (e.g. bli and individual files) don't have 'subpop':
  ### add it in as NAs.
  # if(!'subpop' %in% names(spp_dbf_info)) spp_dbf_info$subpop <- NA
  
  ### some files don't have 'presence' field: if not present or NA,
  ### set to presence = 1.  Some have presence = 0?
  # if(!'presence' %in% names(spp_dbf_info)) spp_dbf_info$presence <- 1
  
  # spp_dbf_info <- spp_dbf_info %>%
  #   mutate(presence = ifelse(presence == 0, 1, presence))
  
  return(spp_dbf_info)
}

shps_df <- lapply(shps_dbfs, get_dbf) %>%
  setNames(shps_dbfs) %>%
  bind_rows(.id = 'dbf_file') %>%
  dplyr::select(dbf_file, iucn_sid, sciname, presence, subpop) %>%
  filter(iucn_sid %in% marine_spp$iucn_sid)

shp_subpops <- shps_df %>%
  dplyr::select(shp_iucn_sid = iucn_sid, sciname, shp_subpop = subpop) %>%
  filter(!is.na(shp_subpop)) %>%
  distinct() %>%
  mutate(shp_subpop_clean = str_replace(tolower(shp_subpop), ' subpopulation| ocean', ''),
         shp_subpop_clean = str_replace_all(shp_subpop_clean, '[^a-z]+', ' '),
         shp_subpop_clean = str_trim(shp_subpop_clean),
         shp_subpop_clean = str_replace(shp_subpop_clean, 'noth', 'north'))
           ### fix typo

api_subpops <- api_spp %>%
  dplyr::select(api_iucn_sid = iucn_sid, sciname, api_subpop = population) %>%
  filter(!is.na(api_subpop)) %>%
  filter(sciname %in% shp_subpops$sciname) %>%
  distinct() %>%
  mutate(api_subpop_clean = str_replace_all(tolower(api_subpop), ' subpopulation| ocean', ''),
         api_subpop_clean = str_replace_all(api_subpop_clean, '[^a-z]+', ' ') %>% str_trim)

subpops_match_raw <- shp_subpops %>%
  full_join(api_subpops, by = 'sciname') %>%
  group_by(api_iucn_sid) %>%
  mutate(subpop_match = str_detect(shp_subpop_clean, api_subpop_clean),
         n_match = sum(subpop_match)) %>%
  filter(subpop_match | sum(subpop_match) == 0) %>%
  ungroup()

caretta_subpops <- subpops_match_raw %>%
  filter(shp_iucn_sid == 3897) %>%
  mutate(api_subpop_single = str_split(api_subpop_clean, ' ')) %>%
  unnest(api_subpop_single) %>%
  group_by(api_subpop_clean, shp_subpop_clean) %>%
  mutate(match = str_detect(shp_subpop_clean, api_subpop_single),
         n_match = sum(match),
         n_words = n()) %>%
  filter(sum(match) == n()) %>%
  ungroup() %>%
  select(shp_iucn_sid, sciname, shp_subpop, api_iucn_sid, api_subpop) %>%
  distinct()

subpops_match <- subpops_match_raw %>%
  filter(shp_iucn_sid != 3897 & subpop_match == TRUE) %>%
  select(shp_iucn_sid, sciname, shp_subpop, api_iucn_sid, api_subpop) %>%
  bind_rows(caretta_subpops)

### Check for missed matches:

# api_subpops$api_iucn_sid[!api_subpops$api_iucn_sid %in% subpops_match$api_iucn_sid]
# 16369383 156206333 181208820 189865869
### * API subpop, not matched: 16369383 Tursiops truncatus 
###     * Mediterranean subpopulation
###     * OK - no polygon for this subpopulation.  Regional assessment?
### * API subpop, not matched: 156206333 Delphinus delphis
###     * Gulf of Corinth subpopulation
### * API subpop, not matched: 181208820 Tursiops truncatus
###     * Gulf of Ambracia subpopulation
### * API subpop, not matched: 189865869 Delphinus delphis
###     * Inner Mediterranean subpopulation

# shp_subpops[!shp_subpops$shp_subpop %in% subpops_match$shp_subpop, ]

###  16208224	Balaenoptera physalus Mediterranean subpopulation	
### 198785664	Globicephala melas Inner Mediterranean subpopulation	
### 198787290	Globicephala melas Strait of Gibraltar subpopulation	
###      6494	Dermochelys coriacea	
###  16285718	Chelonia mydas Hawaiian subpopulation
### * Shapefile subpop, not matched: 16285718 Chelonia mydas Hawaiian subpopulation
###     * OK: the shapefile ID matches the API ID - will match automatically

subpops_file <- here('int', sprintf('subpops_match_api_to_shp_%s.csv', api_version))
write_csv(subpops_match, subpops_file)

```

### get info on comprehensive assessments

Not all maps are for species that are included in comprehensively assessed taxonomic groups.  Get info on comprehensive assessment status for species from the API.

``` {r get comp assess status}
comp_file <- here('_data/iucn_spp', sprintf('iucn_comp_assessed_%s.csv', api_version))
if(!file.exists(comp_file)) {
  comp_gps_url <- 'http://apiv3.iucnredlist.org/api/v3/comp-group/list?token=%s'
  comp_gps <- fromJSON(sprintf(comp_gps_url, api_key)) %>%
    .$result
  
  comp_spp_url <- 'http://apiv3.iucnredlist.org/api/v3/comp-group/getspecies/%s?token=%s'
  comp_spp <- mc_get_from_api(comp_spp_url, comp_gps$group_name, api_key, 
                              add_param_val = TRUE) %>%
    select(iucn_sid = taxonid, sciname = scientific_name, assess_gp = param_val)
  
  ### NOTE: sygnathiformes and clupeiformes not included here? wtf
  ### add them in:
  spp_info <- read_csv(here_anx('iucn_spp',
                                sprintf('spp_info_from_api_%s.csv', api_version)))
  syngnathids <- spp_info %>%
    filter(tolower(family) %in% c('syngnathidae', 'aulostomidae', 
                                  'centriscidae', 'fistulariidae', 
                                  'pegasidae', 'solenostomidae')) %>%
    mutate(assess_gp = 'syngnathiformes')
  clupeids <- spp_info %>%
    filter(tolower(order) == 'clupeiformes') %>%
    mutate(assess_gp = 'clupeiformes')
  ### NOTE: turtles also not included here
  turtles <- c('cheloniidae', 'dermochelyidae')
  turtles <- spp_info %>%
    filter(tolower(family) %in% turtles) %>%
    mutate(assess_gp = 'sea_turtles')
    
  comp_spp <- bind_rows(comp_spp, clupeids, syngnathids, turtles) %>%
    select(iucn_sid, sciname, assess_gp) %>%
    distinct()
  
  write_csv(comp_spp, comp_file)
} else {
  # comp_spp <- read_csv(comp_file)
}
```

### Join map list to subpop corrected list; remove duplicated spp IDs

From the species range map shapefiles, pull map info from the .dbf files.  For species listed in multiple files (e.g. sea snakes and reptiles), remove duplicates.  For species with polygons differentiated by subpopulation, adjust the iucn_sid to match the subpop info.

``` {r generate map list from dbfs}

iucn_dbfs  <- list.files(dir_shp, pattern = '\\.dbf$', full.names = TRUE)
bli_dbfs   <- list.files(dir_bli, pattern = '\\.dbf$', full.names = TRUE)

map_files  <- c(iucn_dbfs, bli_dbfs)

# spp_group_names <- map_files %>%
#   basename() %>%
#   str_replace_all('\\.dbf$|_PART_.+', '') %>%
#   unique()

map_info_raw <- lapply(map_files, get_dbf) %>%
  setNames(shps_dbfs) %>%
  bind_rows(.id = 'dbf_file') %>%
  select(shp_iucn_sid = iucn_sid, sciname, presence, subpop, dbf_file)

### Notes on presence, origin, seasonal fields:
### * presence = 5 is extinct; 4 = probably extinct; others are extant-ish or 
###   uncertain. We will drop field and include all polygons for now.
### * origin is native, introduced, etc.  We will drop this field and not
###   worry about origin.
### * seasonal is breeding/non-breeding/passage.  We will drop this field
###   and not worry about seasonality.

# subspp <- map_info %>%
#   filter(!is.na(subspecies)) %>%
#   select(id_no, code) %>%
#   distinct()
### No subspecies vary by IUCN code; drop field for simplicity
### and include all at species level

subpops_match <- read_csv(here('int', sprintf('subpops_match_api_to_shp_%s.csv', api_version)))

map_info_add_subpops <- map_info_raw %>%
  full_join(subpops_match, by = c('shp_iucn_sid', 'sciname', 'subpop' = 'shp_subpop')) %>%
  mutate(iucn_sid = ifelse(is.na(api_iucn_sid), shp_iucn_sid, api_iucn_sid)) %>%
  select(shp_iucn_sid, iucn_sid, sciname, presence, subpop, dbf_file) %>%
  distinct()

### filter to marine habitat species (to drop terrestrial reptiles e.g.)
marine_spp_ids <- read_csv(here('_data/iucn_spp', sprintf('spp_marine_from_api_%s.csv',
                                                 api_version))) %>%
  select(iucn_sid, max_depth)

### Filter clupeiformes from Caspian sea (found because they result
### in zero-cell rangefiles despite large polygons):
caspian_spp <- c(135548, 135707, 98468449, 98468550, 98470302, 98471289, 98471433)

marine_map_info <- map_info_add_subpops %>%
  inner_join(marine_spp_ids, by = 'iucn_sid') %>%
  filter(!iucn_sid %in% caspian_spp)

write_csv(marine_map_info, here('_data/iucn_spp', sprintf('spp_marine_maps_%s.csv', api_version)))

```

## Table of species to include

To this point we have created a list of all marine species with maps in the shapefiles from IUCN.  Here we will generate a table of the subset of marine species considered in this study.  Included species meet the following criteria:

* species has a polygon feature in one of the shapefiles
* species is listed in a group with comprehensively assessed status (incl manually adjusted ones)
* species extinction risk is threatened or near threatened (i.e. not Least Concern, Extinct, or Data Deficient)

Note that when processing these down, we will focus on threatened marine species with sensitivity to the particular stressors; additionally, some species listed here may have maps that don't overlap with ocean area, creating a degenerate range map.

``` {r generate table of species}

marine_map_info <- read_csv(here('_data/iucn_spp',
    sprintf('spp_marine_maps_%s.csv', api_version)),
    col_types = cols('subpop' = 'c')) %>%
  filter(presence != 5) %>%
  select(-presence) %>%
  distinct()

taxa_info <- read_csv(here_anx('iucn_spp', 
                               sprintf('spp_info_from_api_%s.csv', api_version))) %>%
  select(iucn_sid, kingdom:family)

comp_info <- read_csv(comp_file)

cat_info <- read_csv(here('_data/iucn_spp', 
                          sprintf('iucn_marine_spp_info_%s.csv', api_version))) %>%
  select(iucn_sid, cat_score) %>%
  filter(!is.na(cat_score) & !cat_score %in% c(0, 1))

included_maps <- marine_map_info %>%
  filter(iucn_sid %in% comp_info$iucn_sid) %>%
  filter(iucn_sid %in% cat_info$iucn_sid) %>%
  select(iucn_sid, shp_iucn_sid, subpop, sciname, dbf_file) %>%
  left_join(taxa_info, by = 'iucn_sid') %>%
  mutate(spp_gp = basename(dbf_file),
         spp_gp = spp_gp %>% tolower() %>% str_replace('.dbf|_part.+', '')) %>%
  mutate(kingdom = tools::toTitleCase(tolower(kingdom)),
         phylum  = tools::toTitleCase(tolower(phylum)),
         order   = tools::toTitleCase(tolower(order)),
         class   = tools::toTitleCase(tolower(class)),
         family  = tools::toTitleCase(tolower(family)))

all_in_the_fam <- included_maps %>%
  group_by(kingdom, phylum, class, order, family, spp_gp) %>%
  summarize(species = length(unique(shp_iucn_sid)),
            subpops = sum(!is.na(subpop)))  

basic_sum <- included_maps %>%
  filter(!str_detect(tolower(class), 'actinopt|mammal|aves|rept')) %>%
  group_by(kingdom, phylum, class, spp_gp) %>%
  summarize(species = length(unique(shp_iucn_sid)),
            subpops = sum(!is.na(subpop)))

nonfish_sum <- included_maps %>%
  filter(str_detect(tolower(class), 'mammal|aves|rept')) %>%
  group_by(kingdom, phylum, class, order, spp_gp) %>%
  summarize(species = length(unique(shp_iucn_sid)),
            subpops = sum(!is.na(subpop)))

fish_sum <- included_maps %>%
  filter(str_detect(tolower(class), 'actinopt')) %>%
  group_by(kingdom, phylum, class, order, family, spp_gp) %>%
  summarize(species = length(unique(shp_iucn_sid)),
            subpops = sum(!is.na(subpop)))

all_sum <- bind_rows(fish_sum, nonfish_sum, basic_sum) %>%
  group_by(spp_gp, kingdom, phylum, class, order) %>%
  summarize(species = sum(species),
            subpops = sum(subpops, na.rm = TRUE)) %>%
  ungroup() %>%
  mutate(subpops = ifelse(subpops == 0, NA, subpops))

write_csv(all_sum, here('_output/iucn_spp', 'spp_included_summary.csv'))
write_csv(all_in_the_fam, here('_output/iucn_spp', 'spp_included_by_family.csv'))

knitr::kable(all_sum)
```
